Update for Nuget.org Submission
Changed csproj to prepare for Nuget.org submission - this includes
a change to the package id, as we need to mirror the Nuget naming
rules for Microsoft packages.

This also includes an copy of the team logo in the assets directory,
so it can be referenced in the nuget package.